AWSManagementAgents/requirements.txt
langchain
langchain-groq
langchain-core
boto3
python-dotenv
pydantic
